backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
python-dotenv==1.0.1
